Project/Project/requirements.txt
flask==2.3.3
pyserial==3.5
jsonschema==4.19.2
python-dateutil==2.8.2
requests==2.31.0
flask-cors==4.0.0
